Add getPostDetailBySlug method to PostController and related changes
This commit adds a new method, getPostDetailBySlug, to the PostController class. This method retrieves post details based on the slug parameter and includes the necessary data for rendering the post detail view. Additionally, the necessary changes have been made to the PostModel class to support this new method. The view files for the post detail page have also been added.